+++ docs/Changelog.md
@@ -26,6 +26,7 @@
     * Added support for PerformSingleExecution action in DeploymentManager and in GUI
     * Improved scenario state management to include information about current and deployed versions and allow more customization
 * [#7184](https://github.com/TouK/nussknacker/pull/7184) Improve Nu Designer API notifications endpoint, to include events related to currently displayed scenario
+* [#7323](https://github.com/TouK/nussknacker/pull/7323) Improve Periodic DeploymentManager db queries
 
 ### 1.18.1 (9 December 2024)